First attempt at rearranging main loop
Moved first iteration out before the loop. Moved the model evaluation to
the end of the loop and let it start with the optimisation part.

This isn't quite working perfectly but getting there with which is this
iteration and what is being set up for the next iteration ii+1.

This is in preparation for the trust region algorithm.